add 2HA and stack pooling

+++ layers/eight_mile/tf/layers.py
@@ -3038,7 +3038,7 @@ class SingleHeadReduction(AttentionReduction):
         return x
 
     def _mean_pool(self, x, seq_lengths):
-        return tf.reduce_sum(x, 1) / tf.expand_dims(seq_lengths, -1)
+        return tf.reduce_sum(x, 1) / tf.expand_dims(tf.cast(seq_lengths, tf.float32), -1)
 
     def _max_pool(self, x, _):
         x = tf.reduce_max(x, 1)
